backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
celery[redis]>=5.3.6
redis>=5.0.1
odfpy>=1.4.1
minio>=7.2.3
alembic>=1.13.1
pydantic-settings>=2.1.0
httpx>=0.26.0
pyjwt>=2.8.0
asgiref>=3.7.2
pypdf>=3.17.0
pymupdf>=1.23.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
langgraph>=0.2.0
